Implemented Delete
Yay! Got delete implemented for comments, blog posts, and seeds. Triggers are set in place in the DB so if a post is deleted, its comments will be too, and if a seed is deleted, its respective tag is deleted too. I kept tags in another table, incase we would ever add functionality to add more tags. Just why not I guess. Also added empty message to comments so if there are no comments, it will say so. A lot still needs styling lowkey but that is basically all the functionality