Finished a lot
Modified the homepage so now our photo gallery works and looks great and formatted. I also changed a bit on the get mc page so it is a little more streamlined and looks better in my opinion. Just got rid of the whole second list since it was kinda redundant; put a short list in instead.